+++ infra/bootstrap/gcp/apis.tf
@@ -1,0 +1,14 @@
+# Minimal APIs the bootstrap stack itself needs (WIF + tfstate bucket + project IAM).
+resource "google_project_service" "bootstrap" {
+  for_each = toset([
+    "iam.googleapis.com",
+    "iamcredentials.googleapis.com",
+    "sts.googleapis.com",
+    "cloudresourcemanager.googleapis.com",
+    "storage.googleapis.com",
+    "serviceusage.googleapis.com",
+  ])
+
+  service            = each.value
+  disable_on_destroy = false
+}